requirements.txt
fastapi==0.119.0
pydantic==2.12.2
pydantic-settings==2.11.0
python-decouple==3.8
SQLAlchemy==2.0.42
alembic==1.16.5
psycopg2-binary==2.9.11